+++ NEWS.md
@@ -1,11 +1,20 @@
+# IEATools 0.1.54 (2021-07-09) 
 
-# IEATools 0.1.53 (2021-05-07)
+* Modifications to enable per-machine
+  final-to-useful efficiency data.
+* New tests for new features.
+    * Up to 790 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# IEATools 0.1.53 (2021-06-08)
 
 * The "Oil and gas extraction" activity is split in "Oil extraction",
   which extracts oil products, and "Natural gas extraction", which extracts natural gas.
 * New `split_oil_gas_extraction_eiou()` function splits the EIOU of 
   the "Oil and gas extraction" industry into EIOU for the
   "Oil extraction" and for the "Natural gas extraction" activities.
+
 
 # IEATools 0.1.52 (2021-05-07)